Fix tests

+++ Modules/posixmodule.c
@@ -5540,7 +5540,7 @@ os__path_abspath(PyObject *module, PyObject *path)
 
         cwd_len = wcslen(cwd);
         abs_path_len = cwd_len + 1 + rel_path_len;
-        if (abs_path_len + 1 > PY_SSIZE_T_MAX / sizeof(wchar_t)) {
+        if (abs_path_len + 1 > (Py_ssize_t)PY_SSIZE_T_MAX / sizeof(wchar_t)) {
             return posix_error();
         }
